refactor(app): extract viewmodel and webview files from MainActivity
